Fast + socket.io/requirements.txt
# server.py
uvicorn
fastapi
python-socketio
websockets

# client.py
requests 
websocket-client